MON-234 add script to generate output id

Approved-by: Quentin Manfroi <quentin.manfroi@yahoo.fr>
Approved-by: Jérôme Respaut <shr3ps@gmail.com>
Approved-by: Patrick Decat <patrick.decat@fr.clara.net>
Approved-by: Alex Lemaresquier <alex+bitbucket@lemaresquier.org>

+++ scripts/02_update_readme.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -xeo pipefail
 
 script_dir=$(dirname $0)
 if [[ "$script_dir" == "." ]]; then
@@ -6,6 +7,8 @@ if [[ "$script_dir" == "." ]]; then
 else
   cd "$(dirname $script_dir)"
 fi
+
+PATTERN_DOC="Related documentation"
 
 for dir in $(find -mindepth 2 -name README.md); do
     cd $(dirname $dir)
@@ -17,7 +20,11 @@ for dir in $(find -mindepth 2 -name README.md); do
     sed -i '/Inputs/a ------' README.md
     sed -i 's/## Outputs/Outputs/g' README.md
     sed -i '/Outputs/a -------' README.md
-    grep -Pzo --color=never '.*Related documentation(.*\n)*' README.md.bak | head -n -1 >> README.md
+    if ! grep "${PATTERN_DOC}" README.md.bak; then
+        echo "Error: missing documentation section in README"
+        exit 1
+    fi
+    grep -Pzo --color=never ".*${PATTERN_DOC}(.*\n)*" README.md.bak | head -n -1 >> README.md
     dos2unix README.md
     rm README.md.bak
     cd - >> /dev/null
